requirements.txt
annotated-types==0.7.0
anyio==4.4.0
camelot-py==0.9.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
distro==1.9.0
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
JPype1==1.5.0
numpy==2.0.0
openai==1.35.13
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyPDF2==1.26.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sniffio==1.3.1
tabula-py==2.9.3
tenacity==8.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
